Avoid running header parsing code when no longer in the header
The change looks huge due to indentation changes, however the only real change is to
split 'if ((ofx_start == false) && (<OFX> tag detection)' into two separate if clauses.
The 'else' part s now really only executed when not in the header.
In the original code it would also be executed when in the OFX body for each line
except the <OFX>/<OFC> ones although there's no point any more to do so.