Rewrite sanitize_proprietary_tags
Instead of iterating of each character in the input string
actively search for opening and (optional) closing tags.
If the opening tag turns out to be proprietary remove everything
until (and including) it's matching closing tag. In the absence of a matching
closing tag, remove everything until (but not including) the next
tag.