configure.ac: Respect --libdir for OpenSP
When searching for OpenSP, we want to use
the libdir passed in to autotools (--libdir)
to ensure that we find the library
for the correct ABI.

It is possible that we pick up the wrong
copy from e.g. /usr/lib/ where a 32-bit
copy of OpenSP exists
when we're in the middle of a 64-bit build.

Use ${libdir} to ensure we respect
any preferences/information given and
search for OpenSP in the right place.

Bug: https://bugs.gentoo.org/693458
Signed-off-by: Sam James <sam@gentoo.org>