Refactor ofxdate_to_time_t.
Take a const string& to avoid the copy.
Reorder the conditions to reduce nesting and improve flow.
Remove unused variables.
memset the struct tm to 0 instead of to the current local time.
Use std::string::empty() instead of std::string::size() == 0.
In the empty string case just return time(NULL) instead of converting
that to a struct tm and then back to a time_t.
Move declarations of variables to point of use and initialize instead of
assign.
Apply the timezone offset directly to struct tm an call timegm instead
of messing around with the user's timezone and worrying about  whether
it's in DST. checked_mktime is no longer needed.
Rewrite the doxygen comment to be more concise.