CMake: use GNUInstallDirs
This works fine on Windows too and other non-Linux platforms, despite
the name.

It allows overriding directories in the way distributions expect
and is the defacto solution for getting paths right.

After this change, e.g. the pkgconfig file gets installed to
the correct libdir rather than always /usr/lib/pkgconfig (which
would be wrong for 64-bit systems with /usr/lib64, for example).

Signed-off-by: Sam James <sam@gentoo.org>